move register random vars and priors to ModelBuilder
